Edited mixed test case id.
